Change C2 to 470pF
After doing some breadboard testing the carrier frequency of the PWM
signal was too low at only ~12kHz to ~14kHz. The PWM fan spec says this
needs to be nominally 25kHz with a range of 21kHz to 28kHz. By
decreasing the capacitance of the C2 to 470pF from 680pF this should
adjust the frequency to be in this range. Next time I'll do the
arithmetic before ordering parts.